Redirect stdout and stderr to GUI window.
Bad, bad, bad solution thus far.  Needs some way to update concurrently
to the running processes and also differentiate between stdout and
stderr streams (by colour or so) in the output window.

The log window should also be modal to the first window and exceptions
should be shown as (helpful) message windows.

In the meantime the text-mode version should work the same as before.